optimizations and cleanups

+++ src/org/jbox2d/particle/ParticleColor.java
@@ -11,10 +11,10 @@ public class ParticleColor {
   public byte r, g, b, a;
 
   public ParticleColor() {
-    r = (byte) 127;
-    g = (byte) 127;
-    b = (byte) 127;
-    a = (byte) 50;
+    r = 127;
+    g = 127;
+    b = 127;
+    a = 50;
   }
 
   public ParticleColor(byte r, byte g, byte b, byte a) {
